execute invoke if cmd is invoke
