+++ sdk/java/coop-rest/src/main/kotlin/io/civis/blockchain/coop/rest/config/WebFluxConfiguration.kt
@@ -1,0 +1,26 @@
+package io.civis.blockchain.coop.rest.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.reactive.config.CorsRegistry
+import org.springframework.web.reactive.config.EnableWebFlux
+import org.springframework.web.reactive.config.ResourceHandlerRegistry
+import org.springframework.web.reactive.config.WebFluxConfigurer
+
+@Configuration
+@EnableWebFlux
+class WebFluxConfiguration : WebFluxConfigurer {
+
+    override fun addCorsMappings(corsRegistry: CorsRegistry) {
+        corsRegistry.addMapping("/**")
+                .allowedOrigins("*")
+                .maxAge(3600);
+    }
+
+    override fun addResourceHandlers(registry: ResourceHandlerRegistry) {
+        registry.addResourceHandler("/swagger-ui.html**")
+                .addResourceLocations("classpath:/META-INF/resources/")
+
+        registry.addResourceHandler("/webjars/**")
+                .addResourceLocations("classpath:/META-INF/resources/webjars/")
+    }
+}
